Fixed file loading issue and added latest deploy zip

+++ cloud_connector/cloud_connector/Services/GuardianConfigurationService.cs
@@ -50,7 +50,7 @@ namespace CloudConnector.Services
                 folder = await storageLibrary.SaveFolder.GetFolderAsync(directoryName);
             var storageFile = await folder.TryGetItemAsync(configFileName);
 
-            if (_resetConfig)
+            if (_resetConfig && storageFile != null)
             {
                 await _misty.SendDebugMessageAsync("Resetting config...");
                 await storageFile.DeleteAsync();
